pyproject.toml
[tool.poetry]
name = "langchain-rag"
version = "0.1.0"
description = ""
authors = ["in-sukim <rladlstn494@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
langchain = "^0.2.5"
langchain-community = "^0.2.5"
bs4 = "^0.0.2"
sentence-transformers = "^3.0.1"
faiss-cpu = "^1.8.0"
openai = "^1.35.3"
langchainhub = "^0.1.20"
jupyter = "^1.0.0"
ipykernel = "^6.29.5"
langchain-teddynote = "^0.0.7"
langchain-chroma = "^0.1.2"
streamlit = "^1.36.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
